[APP-5042] Convert MessageCard to TS (#1222)
* Convert MessageCard to TS

* Del Props

* Update components/organisms-simple/MessageCard/MessageCard.tsx


* Improves SimpleSummary tests
