+++ components/molecules/CardContentDescription/CardContentDescription.story.jsx
@@ -3,7 +3,7 @@ import {boolean, text} from '@storybook/addon-knobs';
 import {forceReRender, storiesOf} from '@storybook/react';
 import React from 'react';
 import Input from '~components/atoms/Input/Input';
-import SimpleSummary from '~components/molecules/CardSummaries/SimpleSummary/SimpleSummary';
+import SimpleSummary from '~components/molecules/CardSummaries/SimpleSummary/SimpleSummary.tsx';
 import QuestionCard from '~components/organisms-simple/QuestionCard/QuestionCard';
 import CardContentDescription from './CardContentDescription';
 import CardContentDescriptionReadMe from './CardContentDescription.mdx';
